+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added class DatetimeParser, which parses wrongly defined DOMS dates.
+
 
 ## [1.5.5](https://github.com/kb-dk/kb-util/tree/kb-util-1.5.5)
 ### Changed